change compliesWithLimits(str, double) to compliesWithLimits(str, BigDecimal)

+++ catalog/src/main/java/org/killbill/billing/catalog/DefaultUsage.java
@@ -21,6 +21,7 @@ import java.io.Externalizable;
 import java.io.IOException;
 import java.io.ObjectInput;
 import java.io.ObjectOutput;
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 
@@ -180,7 +181,7 @@ public class DefaultUsage extends ValidatingConfig<StandaloneCatalog> implements
     }
 
     @Override
-    public boolean compliesWithLimits(final String unit, final double value) {
+    public boolean compliesWithLimits(final String unit, final BigDecimal value) {
         final Limit limit = findLimit(unit);
         if (limit != null && !limit.compliesWith(value)) {
             return false;
